fix(react): Fix toggle button always showing as Open
Issue: The toggle button was always showing as open grid icon.

RCA: Search operation of grid was being done in the column factory method which had stale values of gridSchemaStore and searchresult as column factory method is only called once as generateColumns is memoised and grid schema is not changed.