fix(react): Add cell data types in column definition providers
+ Removed MUI DataGrid apiRef and context:
Previously, JsocGrid was wrapped in a context containing the MUI DataGrid API ref, and a hook was used to restore the scroll position while navigating to parent grid from subgrid.
But it is no longer needed since the adapters are wrapped inside the `Activity` component.

The context is currently unused within the repository, so we remove it to eliminate the warning.
If there is a future need to expose the API, we need to re-introduce the context alongside proper mounting guards.